LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-type-containment-expected.txt

AA
A

FAIL container-type:inline-size turns on layout containment assert_equals: expected 33 but got 8
FAIL container-type:inline-size turns on inline-size containment assert_equals: expected 0 but got 12
FAIL container-type:size turns on full size containment assert_equals: expected 0 but got 18
PASS container-type:inline/size turns on style containment
